more test fixes

+++ tests/unit_tests/recipes/llama/test_llama31_405b.py
@@ -187,7 +187,7 @@ class TestPretrainConfig:
         assert config.model.pipeline_model_parallel_size == 16
         assert config.model.context_parallel_size == 8
         assert config.model.sequence_parallel is False
-        assert config.model.pipeline_dtype == torch.bfloat16
+        assert config.model.pipeline_dtype == torch.float32
         assert config.model.virtual_pipeline_model_parallel_size == 4
         assert config.model.account_for_embedding_in_pipeline_split is False
         assert config.model.account_for_loss_in_pipeline_split is False
